Foudn out I had put a bunch of ocntrollers in a class file.  MOved those. I can view a list of cases works.
