+++ src/behaviorTest/java/org/erpmicroservices/endpoint/graphql/CucumberSpringBootContext.java
@@ -3,6 +3,7 @@ package org.erpmicroservices.endpoint.graphql;
 import io.cucumber.spring.CucumberContextConfiguration;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.PeopleAndOrganizationsEndpointGraphqlJavaApplication;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.repositories.*;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.repositories.*;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.party.contactmechanism.PartyContactMechanismPurposeRepository;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.*;
 import org.springframework.boot.test.autoconfigure.graphql.tester.AutoConfigureGraphQlTester;
