+++ src/behaviorTest/java/org/erpmicroservices/endpoint/graphql/steps/Common.java
@@ -4,18 +4,16 @@ import io.cucumber.java.AfterAll;
 import io.cucumber.java.Before;
 import io.cucumber.java.BeforeAll;
 import io.cucumber.java.en.Given;
-import io.cucumber.spring.CucumberContextConfiguration;
 import org.erpmicroservices.endpoint.graphql.CucumberSpringBootContext;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.model.CommunicationEventStatusType;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.repositories.*;
-import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.CaseStatusType;
-import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.CaseType;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseStatusType;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseType;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.repositories.*;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.party.contactmechanism.PartyContactMechanismPurposeRepository;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.party.role.PartyRoleType;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.party.type.PartyType;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.*;
-import org.springframework.boot.test.autoconfigure.graphql.tester.AutoConfigureGraphQlTester;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.graphql.test.tester.GraphQlTester;
