+++ src/main/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/controllers/CaseController.java
@@ -1,7 +1,10 @@
-package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
+package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.controllers;
 
 import graphql.relay.Edge;
 import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.model.CommunicationEvent;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.model.NewCommunicationEvent;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.repositories.CommunicationEventRepository;
@@ -9,7 +12,12 @@ import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicatio
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.communicationevent.repositories.CommunicationEventTypeRepository;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.dto.Cursor;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.dto.PageInfo;
-import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.*;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.graphql.*;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.*;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.repositories.*;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.ContactMechanismTypeRepository;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.PartyRelationshipRepository;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories.PartyRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -24,12 +32,11 @@ import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static java.lang.String.valueOf;
 import static org.erpmicroservices.peopleandorganizations.endpoint.graphql.misc.Utils.pageInfoToPageable;
-
 
 @Controller
 public class CaseController {
-
 	Logger logger = LoggerFactory.getLogger(CaseController.class);
 	private final CaseRepository caseRepository;
 
@@ -50,11 +57,11 @@ public class CaseController {
 	private final PartyRepository partyRepository;
 
 	public CaseController(final CaseRepository caseRepository, final CaseRoleRepository caseRoleRepository,
-	                      final CaseTypeRepository caseTypeRepository,
-	                      final CaseStatusTypeRepository caseStatusTypeRepository,
-	                      final CaseRoleTypeRepository caseRoleTypeRepository,
-	                      final CommunicationEventRepository communicationEventRepository,
-	                      final CommunicationEventStatusTypeRepository communicationEventStatusTypeRepository, final CommunicationEventTypeRepository communicationEventTypeRepository, final ContactMechanismTypeRepository contactMechanismTypeRepository, final PartyRelationshipRepository partyRelationshipRepository, final PartyRepository partyRepository) {
+						  final CaseTypeRepository caseTypeRepository,
+						  final CaseStatusTypeRepository caseStatusTypeRepository,
+						  final CaseRoleTypeRepository caseRoleTypeRepository,
+						  final CommunicationEventRepository communicationEventRepository,
+						  final CommunicationEventStatusTypeRepository communicationEventStatusTypeRepository, final CommunicationEventTypeRepository communicationEventTypeRepository, final ContactMechanismTypeRepository contactMechanismTypeRepository, final PartyRelationshipRepository partyRelationshipRepository, final PartyRepository partyRepository) {
 		this.caseRepository = caseRepository;
 		this.caseRoleRepository = caseRoleRepository;
 		this.caseTypeRepository = caseTypeRepository;
@@ -67,22 +74,6 @@ public class CaseController {
 		this.partyRelationshipRepository = partyRelationshipRepository;
 		this.partyRepository = partyRepository;
 	}
-
-//	@QueryMapping
-//	public CaseConnection cases(@Argument Integer first, @Argument String after, @Argument Integer last, @Argument String before) {
-//		final Page<Case> all = caseRepository.findAll(pageInfoToPageable(pageInfo));
-//		logger.trace("Total Elements: %d, Total Pages: %d".formatted(all.getTotalElements(), all.getTotalPages()));
-//		final List<Edge<Case>> caseEdges = all.get()
-//				                                   .map(kase -> CaseEdge.builder()
-//						                                                .node(kase)
-//						                                                .connectionCursor(Cursor.builder().value(String.valueOf(kase.getId().hashCode())).build())
-//						                                                .build())
-//				                                   .collect(Collectors.toList());
-//		return CaseConnection.builder()
-//				       .edges(caseEdges)
-//				       .pageInfo(pageInfo)
-//				       .build();
-//	}
 
 	@SchemaMapping
 	public CaseType caseType(@NotNull Case kase) {
@@ -98,57 +89,57 @@ public class CaseController {
 	public CaseRoleConnection roles(@Argument PageInfo pageInfo, Case kase) {
 		final Page<CaseRole> byKase_id = caseRoleRepository.findByCaseId(kase.getId(), pageInfoToPageable(pageInfo));
 		final List<Edge<CaseRole>> caseRoleEdges = byKase_id.get()
-				                                           .map(caseRole -> CaseRoleEdge.builder()
-						                                                            .node(caseRole)
-						                                                            .cursor(Cursor.builder().value(String.valueOf(caseRole.getId().hashCode())).build())
-						                                                            .build())
-				                                           .collect(Collectors.toList());
+				.map(caseRole -> CaseRoleEdge.builder()
+						.node(caseRole)
+						.cursor(Cursor.builder().value(String.valueOf(caseRole.getId().hashCode())).build())
+						.build())
+				.collect(Collectors.toList());
 		return CaseRoleConnection.builder()
-				       .edges(caseRoleEdges)
-				       .pageInfo(pageInfo)
-				       .build();
+				.edges(caseRoleEdges)
+				.pageInfo(pageInfo)
+				.build();
 	}
 
 	@MutationMapping
 	public Case createCase(@Argument NewCase newCase) {
 		return caseRepository.save(Case.builder()
-				                           .caseStatusTypeId(newCase.getCaseStatusTypeId())
-				                           .caseTypeId(newCase.getCaseTypeId())
-				                           .description(newCase.getDescription())
-				                           .startedAt(newCase.getStartedAt())
-				                           .build());
+				.caseStatusTypeId(newCase.getCaseStatusTypeId())
+				.caseTypeId(newCase.getCaseTypeId())
+				.description(newCase.getDescription())
+				.startedAt(newCase.getStartedAt())
+				.build());
 	}
 
 	@MutationMapping
 	public Case addCaseRole(@Argument NewCaseRole newCaseRole) {
 		caseRoleRepository.save(CaseRole.builder()
-				                        .caseId(newCaseRole.getCaseId())
-				                        .caseRoleTypeId(newCaseRole.getCaseRoleTypeId())
-				                        .partyId(newCaseRole.getPartyId())
-				                        .fromDate(newCaseRole.getFromDate())
-				                        .build());
+				.caseId(newCaseRole.getCaseId())
+				.caseRoleTypeId(newCaseRole.getCaseRoleTypeId())
+				.partyId(newCaseRole.getPartyId())
+				.fromDate(newCaseRole.getFromDate())
+				.build());
 		return caseRepository.findById(newCaseRole.getCaseId()).get();
 	}
 
 	@MutationMapping
 	public CaseRole expireCaseRole(@Argument UUID caseId, @Argument UUID caseRoleId) {
 		return caseRoleRepository.findById(caseRoleId)
-				       .map(kaseRole -> {
-					       kaseRole.setThruDate(LocalDate.now());
-					       return caseRoleRepository.save(kaseRole);
-				       }).orElseThrow();
+				.map(kaseRole -> {
+					kaseRole.setThruDate(LocalDate.now());
+					return caseRoleRepository.save(kaseRole);
+				}).orElseThrow();
 	}
 
 	@MutationMapping
 	public CommunicationEvent addCommunicationEventToCase(@Argument UUID caseId, @Argument NewCommunicationEvent newCommunicationEvent) {
 		return communicationEventRepository.save(CommunicationEvent.builder()
-				                                         .caseId(caseId)
-				                                         .communicationEventStatusTypeId(newCommunicationEvent.getCommunicationEventStatusTypeId())
-				                                         .communicationEventTypeId(newCommunicationEvent.getCommunicationEventTypeId())
-				                                         .contactMechanismTypeId(newCommunicationEvent.getContactMechanismTypeId())
-				                                         .note(newCommunicationEvent.getNote())
-				                                         .partyRelationshipId(newCommunicationEvent.getPartyRelationshipId())
-				                                         .started(newCommunicationEvent.getStarted())
-				                                         .build());
+				.caseId(caseId)
+				.communicationEventStatusTypeId(newCommunicationEvent.getCommunicationEventStatusTypeId())
+				.communicationEventTypeId(newCommunicationEvent.getCommunicationEventTypeId())
+				.contactMechanismTypeId(newCommunicationEvent.getContactMechanismTypeId())
+				.note(newCommunicationEvent.getNote())
+				.partyRelationshipId(newCommunicationEvent.getPartyRelationshipId())
+				.started(newCommunicationEvent.getStarted())
+				.build());
 	}
 }
