+++ src/main/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/graphql/CaseEdge.java
@@ -1,21 +1,22 @@
-package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
+package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.graphql;
 
 import graphql.relay.ConnectionCursor;
 import graphql.relay.Edge;
 import lombok.Builder;
 import lombok.Data;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.Case;
 
 @Data
 @Builder
-public class CaseEdge implements Edge<Case> {
-	private Case node;
+public class CaseEdge implements Edge<CaseNode> {
+	private CaseNode node;
 	private ConnectionCursor connectionCursor;
 
 	/**
 	 * @return the node of data that this edge represents
 	 */
 	@Override
-	public Case getNode() {
+	public CaseNode getNode() {
 		return node;
 	}
