+++ src/main/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/models/CaseRoleType.java
@@ -1,12 +1,9 @@
-package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
+package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models;
 
 import jakarta.persistence.Entity;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.models.AbstractType;
 
 import java.util.UUID;
@@ -14,10 +11,11 @@ import java.util.UUID;
 @Entity
 @Data
 @EqualsAndHashCode(callSuper = true)
+@ToString(callSuper = true)
 @NoArgsConstructor
-public class CaseStatusType extends AbstractType<CaseStatusType> {
+public class CaseRoleType extends AbstractType<CaseRoleType> {
 	@Builder
-	public CaseStatusType(@NotNull final UUID id, @NotBlank final String description, final UUID parentId) {
+	public CaseRoleType(@NotNull final UUID id, @NotBlank final String description, final UUID parentId) {
 		super(id, description, parentId);
 	}
 }
