+++ src/main/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/models/CaseStatusType.java
@@ -1,9 +1,12 @@
-package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
+package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models;
 
 import jakarta.persistence.Entity;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import lombok.*;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import org.erpmicroservices.peopleandorganizations.endpoint.graphql.models.AbstractType;
 
 import java.util.UUID;
@@ -11,11 +14,10 @@ import java.util.UUID;
 @Entity
 @Data
 @EqualsAndHashCode(callSuper = true)
-@ToString(callSuper = true)
 @NoArgsConstructor
-public class CaseRoleType extends AbstractType<CaseRoleType> {
+public class CaseStatusType extends AbstractType<CaseStatusType> {
 	@Builder
-	public CaseRoleType(@NotNull final UUID id, @NotBlank final String description, final UUID parentId) {
+	public CaseStatusType(@NotNull final UUID id, @NotBlank final String description, final UUID parentId) {
 		super(id, description, parentId);
 	}
 }
