+++ src/main/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/repositories/CaseRoleRepository.java
@@ -1,6 +1,6 @@
-package org.erpmicroservices.peopleandorganizations.endpoint.graphql.repositories;
+package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.repositories;
 
-import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.CaseRole;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseRole;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
