+++ src/test/java/org/erpmicroservices/peopleandorganizations/endpoint/builders/CaseRoleTypeTestDataBuilder.java
@@ -1,6 +1,6 @@
 package org.erpmicroservices.peopleandorganizations.endpoint.builders;
 
-import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.CaseRoleType;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseRoleType;
 
 import java.util.UUID;
