+++ src/test/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/CaseRoleTypesQueryTest.java
@@ -1,5 +1,7 @@
 package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
 
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.graphql.CaseRoleTypeEdge;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseRoleType;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
