+++ src/test/java/org/erpmicroservices/peopleandorganizations/endpoint/graphql/kase/CaseTypeQueryTest.java
@@ -1,5 +1,7 @@
 package org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase;
 
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.graphql.CaseTypeEdge;
+import org.erpmicroservices.peopleandorganizations.endpoint.graphql.kase.models.CaseType;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
